+++ src/main/java/seedu/address/logic/commands/eventCommands/AddEventCommand.java
@@ -1,0 +1,5 @@
+package seedu.address.logic.commands.eventCommands;
+
+public class AddEventCommand {
+    
+}
