+++ src/main/java/seedu/address/logic/commands/personCommands/DeleteCommand.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.personCommands;
 
 import static java.util.Objects.requireNonNull;
