+++ src/main/java/seedu/address/logic/commands/personCommands/HelpCommand.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.personCommands;
 
 import seedu.address.model.Model;
