+++ src/main/java/seedu/address/logic/parser/AddressBookParser.java
@@ -8,16 +8,16 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 import seedu.address.commons.core.LogsCenter;
-import seedu.address.logic.commands.AddCommand;
-import seedu.address.logic.commands.ClearCommand;
-import seedu.address.logic.commands.Command;
-import seedu.address.logic.commands.DeleteCommand;
-import seedu.address.logic.commands.EditCommand;
-import seedu.address.logic.commands.ExitCommand;
-import seedu.address.logic.commands.FindCommand;
-import seedu.address.logic.commands.HelpCommand;
-import seedu.address.logic.commands.ListCommand;
-import seedu.address.logic.commands.SearchCommand;
+import seedu.address.logic.commands.personCommands.AddCommand;
+import seedu.address.logic.commands.personCommands.ClearCommand;
+import seedu.address.logic.commands.personCommands.Command;
+import seedu.address.logic.commands.personCommands.DeleteCommand;
+import seedu.address.logic.commands.personCommands.EditCommand;
+import seedu.address.logic.commands.personCommands.ExitCommand;
+import seedu.address.logic.commands.personCommands.FindCommand;
+import seedu.address.logic.commands.personCommands.HelpCommand;
+import seedu.address.logic.commands.personCommands.ListCommand;
+import seedu.address.logic.commands.personCommands.SearchCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 
 /**
