+++ src/main/java/seedu/address/logic/parser/DeleteCommandParser.java
@@ -3,7 +3,7 @@ package seedu.address.logic.parser;
 import static seedu.address.logic.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.DeleteCommand;
+import seedu.address.logic.commands.personCommands.DeleteCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 
 /**
