+++ src/main/java/seedu/address/logic/parser/SearchCommandParser.java
@@ -4,7 +4,7 @@ import static seedu.address.logic.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 
 import java.util.Arrays;
 
-import seedu.address.logic.commands.SearchCommand;
+import seedu.address.logic.commands.personCommands.SearchCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.types.common.TagContainsKeywordsPredicate;
