+++ src/main/java/seedu/address/ui/CommandBox.java
@@ -4,8 +4,8 @@ import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
 import javafx.scene.control.TextField;
 import javafx.scene.layout.Region;
-import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
+import seedu.address.logic.commands.personCommands.CommandResult;
 import seedu.address.logic.parser.exceptions.ParseException;
 
 /**
