+++ src/main/java/seedu/address/ui/MainWindow.java
@@ -13,8 +13,8 @@ import javafx.stage.Stage;
 import seedu.address.commons.core.GuiSettings;
 import seedu.address.commons.core.LogsCenter;
 import seedu.address.logic.Logic;
-import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
+import seedu.address.logic.commands.personCommands.CommandResult;
 import seedu.address.logic.parser.exceptions.ParseException;
 
 /**
