+++ src/test/java/seedu/address/logic/commands/AddCommandTest.java
@@ -18,6 +18,8 @@ import javafx.collections.ObservableList;
 import seedu.address.commons.core.GuiSettings;
 import seedu.address.logic.Messages;
 import seedu.address.logic.commands.exceptions.CommandException;
+import seedu.address.logic.commands.personCommands.AddCommand;
+import seedu.address.logic.commands.personCommands.CommandResult;
 import seedu.address.model.AddressBook;
 import seedu.address.model.Model;
 import seedu.address.model.ReadOnlyAddressBook;
