+++ src/test/java/seedu/address/logic/commands/CommandTestUtil.java
@@ -15,6 +15,9 @@ import java.util.List;
 
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.exceptions.CommandException;
+import seedu.address.logic.commands.personCommands.Command;
+import seedu.address.logic.commands.personCommands.CommandResult;
+import seedu.address.logic.commands.personCommands.EditCommand;
 import seedu.address.model.AddressBook;
 import seedu.address.model.Model;
 import seedu.address.model.types.common.NameContainsKeywordsPredicate;
