+++ src/test/java/seedu/address/logic/commands/EditCommandTest.java
@@ -19,7 +19,9 @@ import org.junit.jupiter.api.Test;
 
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.Messages;
-import seedu.address.logic.commands.EditCommand.EditPersonDescriptor;
+import seedu.address.logic.commands.personCommands.ClearCommand;
+import seedu.address.logic.commands.personCommands.EditCommand;
+import seedu.address.logic.commands.personCommands.EditCommand.EditPersonDescriptor;
 import seedu.address.model.AddressBook;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
