+++ src/test/java/seedu/address/logic/parser/CommandParserTestUtil.java
@@ -2,7 +2,7 @@ package seedu.address.logic.parser;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import seedu.address.logic.commands.Command;
+import seedu.address.logic.commands.personCommands.Command;
 import seedu.address.logic.parser.exceptions.ParseException;
 
 /**
