+++ src/test/java/seedu/address/logic/parser/EditCommandParserTest.java
@@ -36,8 +36,8 @@ import org.junit.jupiter.api.Test;
 
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.Messages;
-import seedu.address.logic.commands.EditCommand;
-import seedu.address.logic.commands.EditCommand.EditPersonDescriptor;
+import seedu.address.logic.commands.personCommands.EditCommand;
+import seedu.address.logic.commands.personCommands.EditCommand.EditPersonDescriptor;
 import seedu.address.model.tag.Tag;
 import seedu.address.model.types.common.Address;
 import seedu.address.model.types.common.Email;
