+++ src/test/java/seedu/address/logic/parser/FindCommandParserTest.java
@@ -8,7 +8,7 @@ import java.util.Arrays;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.logic.commands.FindCommand;
+import seedu.address.logic.commands.personCommands.FindCommand;
 import seedu.address.model.types.common.NameContainsKeywordsPredicate;
 
 public class FindCommandParserTest {
