+++ src/test/java/seedu/address/testutil/PersonUtil.java
@@ -8,8 +8,8 @@ import static seedu.address.logic.parser.CliSyntax.PREFIX_TAG;
 
 import java.util.Set;
 
-import seedu.address.logic.commands.AddCommand;
-import seedu.address.logic.commands.EditCommand.EditPersonDescriptor;
+import seedu.address.logic.commands.personCommands.AddCommand;
+import seedu.address.logic.commands.personCommands.EditCommand.EditPersonDescriptor;
 import seedu.address.model.tag.Tag;
 import seedu.address.model.types.person.Person;
